optimized rest service code

+++ Server/auth/src/resources/user/controller.js
@@ -1,10 +1,10 @@
 import jwt from "jsonwebtoken";
 import bcryptjs from "bcryptjs";
 
-import { tryCatchWrapper } from "../../middlewares/tryCatchWrapper.js";
-import { createCustomError } from "../../errors/customErrors.js";
+import tryCatchWrapper from "../../middlewares/tryCatchWrapper.js";
+import createCustomError from "../../errors/customErrors.js";
 import session from "../../db/helper.js";
-import { generateAccessToken, generateRefreshToken } from "./helper.js";
+import { generateAccessToken, generateRefreshToken, checkLevelPermission} from "./helper.js";
 
 /**
  * @description Create a new user
@@ -25,8 +25,12 @@ export const createUser = tryCatchWrapper(async function (req, res, next) {
   const role = req.body.role;
   const locked = req.body.locked;
 
-  // extract creating user data from token
-  const creatingUserName = req.user.user;
+  // check if req body is present
+  if (newUser === undefined) return next(createCustomError("username is required", 409));
+  if (newUserPwd === undefined) return next(createCustomError("password is required", 409));
+  if (company === undefined) return next(createCustomError("company is required", 409));
+  if (role === undefined) return next(createCustomError("role is required", 409));
+  if (locked === undefined) return next(createCustomError("locked is required", 409));
 
   // sql statements
   const searchUSerSql = "SELECT 1 FROM user.user u WHERE u.name = ? LIMIT 1";
@@ -43,25 +47,26 @@ export const createUser = tryCatchWrapper(async function (req, res, next) {
 
   // check if user already exists
   let [rows] = await session(searchUSerSql, newUser);
-  if (rows.length != 0) return next(createCustomError("User already exists", 409));
+  if (rows.length !== 0) return next(createCustomError("User already exists", 409));
   
   // check if email already exists
   if (email !== undefined) {
     [rows] = await session(searchEmailSql, email);
-    if (rows.length != 0) return next(createCustomError("Email alredy registered", 409));
+    if (rows.length !== 0) return next(createCustomError("Email alredy registered", 409));
   } else email = null;
   
   // check if company exists
   [rows] = await session(searchCompanySql, company);
-  if (rows.length == 0) return next(createCustomError("Company does not exist", 409));
+  if (rows.length === 0) return next(createCustomError("Company does not exist", 409));
 
   // check if role exists
   [rows] = await session(searchRoleSql, role);
-  if (rows.length == 0) return next(createCustomError("Role does not exist", 409));
+  if (rows.length === 0) return next(createCustomError("Role does not exist", 409));
   // retrieve role level and id from new user
-  const newUSerRoleId = rows[0].id;
   const newUserRoleLevel = rows[0].level;
 
+  // extract creating user data from token
+  const creatingUserName = req.user.user;
   // retrieve creating user role level
   [rows] = await session(searchUserRoleSql, creatingUserName);
   const creatingUserRoleLevel = rows[0].level;
@@ -75,7 +80,7 @@ export const createUser = tryCatchWrapper(async function (req, res, next) {
   const hashedPwd = await bcryptjs.hash(newUserPwd, 10);
 
   //insert new user
-  await session(insertSql, [newUser, hashedPwd, email, company, newUSerRoleId, locked]).then((result) => {
+  await session(insertSql, [newUser, hashedPwd, email, company, role, locked]).then((result) => {
     return res.status(201).json({ userId: result[0].insertId });
   });
 });
@@ -87,20 +92,23 @@ export const createUser = tryCatchWrapper(async function (req, res, next) {
  */
 export const enableUser = tryCatchWrapper(async function (req, res, next) {
   // extract data from req parameters
-  const userName = req.params.id;
+  const user = req.params.id;
+
+  // check if user is allowed to enable the user
+  if (!await checkLevelPermission(req.user.user, user)) return next(createCustomError("Cannot enable a user with a higher role than your own", 403));
 
   // sql statements
   const searchSql = "SELECT u.disabled FROM user.user u WHERE u.id = ?  LIMIT 1";
   const updateSql = "UPDATE user.user u SET u.disabled = FALSE WHERE u.id = ?"
 
   // check if user exists
-  const [rows] = await session(searchSql, userName);
-  if (rows.length == 0) return next(createCustomError("User does not exist", 409));
+  const [rows] = await session(searchSql, user);
+  if (rows.length === 0) return next(createCustomError("User does not exist", 409));
   // check if user is already enabled
   if (!rows[0].disabled) return next(createCustomError("User already enabled", 409));
 
   //update user
-  await session(updateSql, userName);
+  await session(updateSql, user);
   return res.status(201).json({ message: "User enabled" });
 });
 
@@ -111,20 +119,23 @@ export const enableUser = tryCatchWrapper(async function (req, res, next) {
  */
 export const disableUser = tryCatchWrapper(async function (req, res, next) {
   // extract data from req parameters
-  const userName = req.params.id;
+  const user = req.params.id;
+
+  // check if user is allowed to enable the user
+  if (!await checkLevelPermission(req.user.user, user)) return next(createCustomError("Cannot disable a user with a higher role than your own", 403));
 
   // sql statements
   const searchSql = "SELECT u.disabled FROM user.user u WHERE u.id = ? LIMIT 1";
   const updateSql = "UPDATE user.user u SET u.disabled = TRUE WHERE u.id = ?"
 
   // check if user exists
-  const [rows] = await session(searchSql, userName);
-  if (rows.length == 0) return next(createCustomError("User does not exist", 409));
+  const [rows] = await session(searchSql, user);
+  if (rows.length === 0) return next(createCustomError("User does not exist", 409));
   // check if user is already disabled
   if (rows[0].disabled) return next(createCustomError("User already disabled", 409));
 
   //update user
-  await session(updateSql, userName);
+  await session(updateSql, user);
   return res.status(201).json({ message: "User disabled" });
 });
 
@@ -138,6 +149,9 @@ export const changePassword = tryCatchWrapper(async function (req, res, next) {
   const newPwd = req.body.password;
   // extract user data from token
   const changingUser = req.user.user;
+
+  // check if req body is present
+  if (newPwd === undefined) return next(createCustomError("password is required", 409));
 
   // sql statements
   const searchPasswordSql = "SELECT u.password FROM user.user u WHERE u.id = ? LIMIT 1";
@@ -181,13 +195,17 @@ export const login = tryCatchWrapper(async function (req, res, next) {
   const user = req.body.user;
   const pwd = req.body.password;
 
+  // check if req body is present
+  if (user === undefined) return next(createCustomError("user is required", 409));
+  if (pwd === undefined) return next(createCustomError("password is required", 409));
+
   // sql statements
   const searchSql = "SELECT u.password, u.role, u.disabled FROM user.user u WHERE u.id = ? LIMIT 1";
   const insertSql = "INSERT INTO user.token VALUES (?, ?)";
 
   // check if user exists
   const [rows] = await session(searchSql, user);
-  if (rows.length == 0) return next(createCustomError("User does not exist", 404));
+  if (rows.length === 0) return next(createCustomError("User does not exist", 404));
   // check if user is disabled
   if (rows[0].disabled) return next(createCustomError("User is disabled", 401));
 
@@ -217,12 +235,12 @@ export const validateToken = (requiredPermission, isMiddleware = true) => {
     // retrieve token from header
     const authHeader = req.headers["authorization"];
 
-    if (authHeader == null) return next(createCustomError("Token not present", 400));
+    if (authHeader === null) return next(createCustomError("Token not present", 400));
   
     const token = authHeader.split(" ")[1];
-    if (token == null) return next(createCustomError("Token not present", 400));
+    if (token === null) return next(createCustomError("Token not present", 400));
 
-    if (requiredPermission == null || requiredPermission == "") return next(createCustomError("Permission not present", 400));
+    if (requiredPermission === null || requiredPermission === "") return next(createCustomError("Permission not present", 400));
 
     // sql statements
     const searchUserSql = "SELECT u.disabled FROM user.user u WHERE u.id = ? LIMIT 1";
@@ -239,7 +257,7 @@ export const validateToken = (requiredPermission, isMiddleware = true) => {
 
     // check if token is valid
     let [rows] = await session(searchTokenSql, token);
-    if (rows.length == 0) return next(createCustomError("Token invalid", 401));
+    if (rows.length === 0) return next(createCustomError("Token invalid", 401));
 
     // verify token
     const payload = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, user) => {
@@ -249,11 +267,11 @@ export const validateToken = (requiredPermission, isMiddleware = true) => {
     });
 
     // extract name and role from token
-    const userName = payload.user;
+    const user = payload.user;
     const userRole = payload.role;
 
     // check if user is disabled
-    [rows] = await session(searchUserSql, userName);
+    [rows] = await session(searchUserSql, user);
     if (rows[0].disabled) return next(createCustomError("User is disabled", 401));
 
     [rows] = await session(searchRolePermissionSql, [userRole, requiredPermission]);
@@ -275,6 +293,9 @@ export const refreshToken = tryCatchWrapper(async function (req, res, next) {
   // extract data from req body
   const refreshTokenOld = req.body.refreshToken;
 
+  // check if req body is present
+  if (refreshTokenOld === undefined) return next(createCustomError("refreshToken is required", 409));
+
   // sql statements
   const searchTokenSql = "SELECT 1 FROM user.token t WHERE t.refresh = ? LIMIT 1";
   const searchUserSql = "SELECT u.disabled FROM user.user u WHERE u.id = ? LIMIT 1";
@@ -282,7 +303,7 @@ export const refreshToken = tryCatchWrapper(async function (req, res, next) {
 
   // check if token is valid
   let [rows] = await session(searchTokenSql, refreshTokenOld);
-  if (rows.length == 0) return next(createCustomError("Refresh token invalid", 400));
+  if (rows.length === 0) return next(createCustomError("Refresh token invalid", 400));
 
   // check if refresh token is expired
   const payload = jwt.verify(refreshTokenOld, process.env.REFRESH_TOKEN_SECRET, (err, user) => {
@@ -291,16 +312,16 @@ export const refreshToken = tryCatchWrapper(async function (req, res, next) {
   });
 
   // retrieve user name from token
-  const userName = payload.user;
+  const user = payload.user;
   const userRole = payload.role;
 
   // check if user is disabled
-  [rows] = await session(searchUserSql, userName);
+  [rows] = await session(searchUserSql, user);
   if (rows[0].disabled) return next(createCustomError("User is disabled", 401));
 
   // generate tokens
-  const accessTokenNew = generateAccessToken(userName, userRole);
-  const refreshTokenNew = generateRefreshToken(userName, userRole);
+  const accessTokenNew = generateAccessToken(user, userRole);
+  const refreshTokenNew = generateRefreshToken(user, userRole);
 
   await session(updateSql, [accessTokenNew, refreshTokenNew, refreshTokenOld]);
   return res.status(201).json({ accessToken: accessTokenNew, refreshToken: refreshTokenNew });
@@ -315,13 +336,16 @@ export const logout = tryCatchWrapper(async function (req, res, next) {
   // extract data from req body
   const refreshToken = req.body.refreshToken;
 
+  // check if req body is present
+  if (refreshToken === undefined) return next(createCustomError("refreshToken is required", 409));
+
   // sql statements
   const searchSql = "SELECT 1 FROM user.token t WHERE t.refresh = ? LIMIT 1";
   const deleteSql = "DELETE FROM user.token t WHERE t.refresh = ?";
 
   // check if token is valid
   const [rows] = await session(searchSql, refreshToken);
-  if (rows.length == 0) return next(createCustomError("Token invalid", 400));
+  if (rows.length === 0) return next(createCustomError("Token invalid", 400));
 
   // check if refresh token is expired
   jwt.verify(refreshToken, process.env.REFRESH_TOKEN_SECRET, (err, user) => {
@@ -341,12 +365,10 @@ export const logout = tryCatchWrapper(async function (req, res, next) {
 export const getIdFromUserName = tryCatchWrapper(async function (req, res, next) {
   // extract data from req parameters
   const name = req.params.name;
-  if (!name) return next(createCustomError("Username is required", 400));
 
   const sql = "SELECT u.id FROM user.user u WHERE u.name = ? LIMIT 1";
   const [rows] = await session(sql, name);
 
-  if (!rows.length) return next(createCustomError("Empty list", 204));
   return res.status(200).json({ user: rows });
 });
 
@@ -358,12 +380,10 @@ export const getIdFromUserName = tryCatchWrapper(async function (req, res, next)
 export const getIdFromCompany = tryCatchWrapper(async function (req, res, next) {
   // extract data from req parameters
   const name = req.params.name;
-  if (!name) return next(createCustomError("Company name is required", 400));
 
   const sql = "SELECT c.id FROM corporation.company c WHERE c.name = ? LIMIT 1";
   const [rows] = await session(sql, name);
 
-  if (!rows.length) return next(createCustomError("Empty list", 204));
   return res.status(200).json({ company: rows });
 });
 
@@ -375,12 +395,10 @@ export const getIdFromCompany = tryCatchWrapper(async function (req, res, next) 
 export const getIdFromRole = tryCatchWrapper(async function (req, res, next) {
   // extract data from req parameters
   const name = req.params.name;
-  if (!name) return next(createCustomError("Role name is required", 400));
 
   const sql = "SELECT r.id FROM privilege.role r WHERE r.name = ? LIMIT 1";
   const [rows] = await session(sql, name);
 
-  if (!rows.length) return next(createCustomError("Empty list", 204));
   return res.status(200).json({ role: rows });
 });
 
@@ -392,7 +410,6 @@ export const getAllCompanies = tryCatchWrapper(async function (req, res, next) {
   const sql = "SELECT c.id FROM corporation.company c";
   const [rows] = await session(sql);
 
-  if (!rows.length) return next(createCustomError("Empty list", 204));
   return res.status(200).json({ companies: rows });
 });
 
@@ -404,6 +421,5 @@ export const getAllRoles = tryCatchWrapper(async function (req, res, next) {
   const sql = "SELECT r.id FROM privilege.role r";
   const [rows] = await session(sql);
 
-  if (!rows.length) return next(createCustomError("Empty list", 204));
   return res.status(200).json({ roles: rows });
 });
